backend/bot/requirements.txt
transformers
torch
python-telegram-bot
requests
cohere